requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-tinymce==3.4.0
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==8.4.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
waitress==2.0.0
whitenoise==5.3.0
